feat(plugin-auto-import): auto-import-plugin support vite mode
